Organização dos arquivos js e dos elementos css
