+++ css/styles.css
@@ -7,7 +7,6 @@ body {
   padding: 0;
 }
 
-/* Container */
 .container {
   max-width: 1200px;
   margin: 0 auto;
@@ -15,7 +14,8 @@ body {
 }
 
 /* Ajuste de imagens e vídeos */
-img, video {
+img,
+video {
   max-width: 100%;
   height: auto;
 }
@@ -52,6 +52,13 @@ img, video {
   color: #ffcc00; /* Rótulos em amarelo pastel */
 }
 
+/* Botões */
+.btn {
+  border-radius: 30px;
+  padding: 10px 20px;
+  font-weight: bold;
+}
+
 .btn-primary {
   background-color: #ffcc00;
   border-color: #ffcc00;
@@ -81,17 +88,28 @@ img, video {
   color: #f0f0f0;
 }
 
-/* Botões */
-.btn {
-  border-radius: 30px;
-  padding: 10px 20px;
-  font-weight: bold;
+.btn-info {
+  max-width: 500px;
+  margin-bottom: 5rem;
+  display: flex;
+  justify-content: center;
+  margin-left: auto;
+  margin-right: auto;
 }
 
 /* Títulos */
-h1, h2 {
+h1,
+h2 {
   color: #f0f0f0;
   font-weight: bold;
+}
+
+/* Títulos das seções */
+h2 {
+  border-bottom: 2px solid #ffcc00;
+  padding-bottom: 10px;
+  margin-top: 40px;
+  color: #ffcc00;
 }
 
 /* Lista de Exercícios */
@@ -139,21 +157,13 @@ h1, h2 {
   color: #ffcc00; /* Amarelo pastel */
 }
 
-/* Títulos das seções */
-h2 {
-  border-bottom: 2px solid #ffcc00;
-  padding-bottom: 10px;
-  margin-top: 40px;
-  color: #ffcc00;
-}
-
 /* Seções */
 .section {
   background-color: #252525;
   border-radius: 8px;
   padding: 20px;
   margin-bottom: 20px;
-  box-shadow: 0 2px 4px rgba(0,0,0,0.2);
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.2);
   transition: transform 0.2s ease;
 }
