+++ index.html
@@ -83,14 +83,16 @@
     </div>
   </div>
 
-  <!-- Importando o jQuery, GSAP, InputMask e Bootstrap JS -->
   <!-- jQuery -->
   <script src="https://code.jquery.com/jquery-3.5.1.min.js"></script>
+
   <!-- GSAP para animações -->
   <script src="https://cdnjs.cloudflare.com/ajax/libs/gsap/3.10.4/gsap.min.js"></script>
   <script src="https://cdnjs.cloudflare.com/ajax/libs/gsap/3.10.4/MorphSVGPlugin.min.js"></script>
+
   <!-- InputMask plugin -->
   <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery.inputmask/5.0.7/jquery.inputmask.min.js"></script>
+
   <!-- Bootstrap JS -->
   <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/js/bootstrap.min.js"></script>
 
@@ -102,7 +104,13 @@
   <!-- Firebase Firestore -->
   <script src="https://www.gstatic.com/firebasejs/9.15.0/firebase-firestore-compat.js"></script>
 
-  <!-- Seu script principal -->
-  <script src="js/app.js"></script>
+  <!-- Scripts JavaScript -->
+  <script src="js/firebase.js"></script>
+  <script src="js/auth.js"></script>
+  <script src="js/utils.js"></script>
+  <script src="js/treino.js"></script>
+  <script src="js/admin.js"></script>
+  <script src="js/exercicio.js"></script>
+  <script src="js/main.js"></script>
 </body>
 </html>
