+++ js/main.js
@@ -1,0 +1,9 @@
+import './utils.js';
+import './exercicio.js';
+import './admin.js';
+import './treino.js';
+import './auth.js';
+import './firebase.js';
+
+// Elementos globais
+export const loader = document.getElementById('loader');
